Tests/LibWeb/Text/expected/css/animated-background-image-timer-doesnt-stop-after-base-change.txt

timer active before base change: true
timer active after removal: true
